+++ plugin-api/src/main/java/org/sonar/api/batch/sensor/rule/AdHocRule.java
@@ -23,6 +23,7 @@ import javax.annotation.CheckForNull;
 import org.sonar.api.batch.rule.Severity;
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.rules.RuleType;
+import org.sonar.api.rules.CleanCodeAttribute;
 
 /**
  * Represents a rule imported from an external rule engine by a {@link Sensor}.
@@ -60,5 +61,13 @@ public interface AdHocRule {
    * Type of the rule.
    */
   RuleType type();
+
+  /**
+   * Clean Code Attribute of the rule according to Clean Code Taxonomy.
+   *
+   * @since 10.1
+   */
+  @CheckForNull
+  CleanCodeAttribute cleanCodeAttribute();
 
 }