+++ plugin-api/src/main/java/org/sonar/api/rules/CleanCodeAttribute.java
@@ -53,4 +53,8 @@ public enum CleanCodeAttribute {
   public CleanCodeAttributeCategory getAttributeCategory() {
     return attributeCategory;
   }
+
+  public static CleanCodeAttribute defaultCleanCodeAttribute() {
+    return CONVENTIONAL;
+  }
 }